initially planned stage is complete. Cloud deployment ready.

+++ BackEnd/src/main/java/com/umesha_g/the_clique_backend/dto/response/ProductCardResponse.java
@@ -10,7 +10,7 @@ public class ProductCardResponse {
     private String name;
     private BigDecimal price;
     private Double rating;
-    private int reviewCount;
+    private long reviewCount;
     private String cardImageUrl;
     private Long purchaseCount;
     private Integer stock;
